add stub of md list toggle to gitcommit, md files
added a todo since i don't know how to do pattern matching on the current
line in vim. but the structure is there to use <S-Tab> to toggle whether
the current line is a list item. eventually want to cycle through list
states in gitignore and markdown files: not a list item, todo list item,
done list item.